// frontend/generated/models/Dog.ts
export interface Dog {
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof Dog
     */
    age: number;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    dob: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    breed: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    gender: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Dog
     */
    photoURLs: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    shelterId: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof Dog
     */
    status: DogStatusEnum;
    color: string;
    size: string;
    shelterName?: string;
    shelterAddress?: string;
    distance?: number;
    shelterEmail?: string;
    shelterContact?: string;
    shelterPostcode?: string;
}


/**
 * @export
 */
export const DogStatusEnum = {
    Available: 'Available',
    Pending: 'Pending Adoption',
    Adopted: 'Adopted'
} as const;
export type DogStatusEnum = typeof DogStatusEnum[keyof typeof DogStatusEnum];